Add test for secret formatting
