+++ elyra/tests/pipeline/test_processor_base.py
@@ -18,9 +18,6 @@ import os
 import pytest
 
 from elyra.pipeline.parser import PipelineParser
-from elyra.pipeline.pipeline import GenericOperation
-from elyra.pipeline.pipeline import KeyValueList
-from elyra.pipeline.pipeline import VolumeMount
 from elyra.pipeline.processor import RuntimePipelineProcessor
 from elyra.tests.pipeline.test_pipeline_parser import _read_pipeline_resource
 
@@ -48,25 +45,3 @@ def sample_metadata():
         "engine": "Argo",
         "tags": [],
     }
-
-
-def test_get_volume_mounts():
-    mounted_volumes = KeyValueList(["/mount/test=rwx-test-claim", "/mount/test_two=second-claim"])
-    component_parameters = {
-        "filename": "pipelines_test_file",
-        "env_vars": [],
-        "runtime_image": "tensorflow/tensorflow:latest",
-        "mounted_volumes": mounted_volumes,
-    }
-    test_operation = GenericOperation(
-        id="this-is-a-test-id",
-        type="execution-node",
-        classifier="execute-notebook-node",
-        name="test",
-        component_params=component_parameters,
-    )
-    parsed_volumes_list = test_operation.get_volume_mounts()
-    assert parsed_volumes_list == [
-        VolumeMount("/mount/test", "rwx-test-claim"),
-        VolumeMount("/mount/test_two", "second-claim"),
-    ]
